fix: adapt useEffect to fetch new endpoint. Add select dropdown for class. Minor optimizations
A change in the API resulted in that the fetch done in the UseEffect got a 400, due to that needed
query params were not provided. A select dropdown was introduced, requiring the user to select their
class.Added memoization for the sorting of masters, to make sure general is at the  very bottom.