feat: refactor webflow with flexbox and Grid for components. Add loading spinner and fix minor bugs
This commit solves issue #34 (which indirectly solves #31, #36, #37).
Added a loading spinner until data is fetched.
Resolved console errors regarding icons and manifests for better SEO
and general good practice.
Added row coloring so that the user knows if it fulfills
the requirements of a master.

TODO: fix repsonsiveness for smaller devices, see #57