refactor: remove usage of SVG as RC. Bump pnpm deps. Refactor missing SC to styles.tsx
Using (SVG) icons as React Components (RC) is outdated and results in errors and slowdown. A long
term solution is to be found, until that, icons are just removed.  Many dependencies were out of
date, and with secuirty in regards, dependeices were bumped to latest versions and tested. Refactor
the last styled components to a separate styles.tsx file for full segretation between components and
styled elements.