fix: Bug when copying studyplan from readonly view (#90)
When copying a studyplan from the readonly view, it did not send an actual request
Which created a new instance of the studyplan. This due to that the savePlan
function in the Context early returned if no unsavedChanges were present.

This fix makes the early return consider if it is a copy operation or not when
saving a plan. Fix #90